requirements.txt
torch
torchvision
pyyaml
easydict
tqdm
matplotlib